pyproject.toml
[project]
name = "project-14"
version = "0.1.0"
description = ""
authors = [
    {name = "TKM-h.almughrabi-c",email = "h.almughrabi-c@takamol.sa"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-mail (>=0.10.0,<1.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
